Rename file to align with class
